LIKA-401: Remove paster commands
