--- ckanext/xroad_integration/commands/__init__.py
@@ -1,6 +1,0 @@
-try:
-    import pkg_resources
-    pkg_resources.declare_namespace(__name__)
-except ImportError:
-    import pkgutil
-    __path__ = pkgutil.extend_path(__path__, __name__)
